refactor: add utils.ts
